backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Grounded Case AI — document understanding and grounded drafting pipeline"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pgvector>=0.3",
    "openai>=1.50",
    "diskcache>=5.6",
    "tenacity>=9.0",
    "python-multipart>=0.0.9",
    "rank-bm25>=0.2.2",
    "pdfplumber>=0.11",
    "pytesseract>=0.3.13",
    "pdf2image>=1.17",
    "pillow>=10.4",
    "minio>=7.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.7",
    "httpx>=0.27",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
